# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Flights API",
        "version": "1.0.0"
    },
    "basePath": "/api/v1/",
    "paths": {
        "/flights": {
            "get": {
                "tags": [
                    "flights"
                ],
                "summary": "Flight List",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "MyHeaderAuthentication": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "flights"
                ],
                "summary": "Add A New Flight",
                "parameters": [
                    {
                        "name": "schedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Field Is Required"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "MyHeaderAuthentication": []
                    }
                ]
            }
        },
        "/flights/create?include={include}&status={status}": {
            "get": {
                "tags": [
                    "flights"
                ],
                "summary": "Include Data & Filter Flight By Status",
                "parameters": [
                    {
                        "name": "include",
                        "in": "path",
                        "description": "Include Data",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Filter by Status",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/flights/{flightNumber}": {
            "get": {
                "tags": [
                    "flights"
                ],
                "summary": "Flight ID",
                "parameters": [
                    {
                        "name": "flightNumber",
                        "in": "path",
                        "description": "Search By Flight Number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "flights"
                ],
                "summary": "Update Flight",
                "parameters": [
                    {
                        "name": "flightNumber",
                        "in": "path",
                        "description": "flightNumber",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Field Is Required"
                    },
                    "505": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "MyHeaderAuthentication": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "flights"
                ],
                "summary": "Remove Flight",
                "operationId": "deleteFlight",
                "parameters": [
                    {
                        "name": "flightNumber",
                        "in": "path",
                        "description": "FlightNumber",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "505": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "MyHeaderAuthentication": []
                    }
                ]
            }
        }
    },
    "definitions": {},
    "securityDefinitions": {
        "MyHeaderAuthentication": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}